Add files via upload
Added a JQuery function to detect which section of the index page the viewport is viewing and playing the corresponding animation for the text and image. Made a JQuery function that activates the detect viewport function when scrolling, when the page is loaded, or when the browser is resized. Added animations for the text for all sections, it moves in from the top and goes from 0 to 1 opacity. Made an animation for all the images that makes the images go from opacity 0 to opacity 1.